Rudimentary kiosk mode and Wayland fixes
The resize bug, whatever it was, caused window to not be drawn
in an Weston/Wayland environment. Changing this to reading
the screen geometry and setting maximized (as many other browsers
seemingly do) works.
Added a rude "kiosk" mode which basically, for now, just sets
fullscreen and disables context menu (right click). Not allowing
downloads and such should be implemented also.
The kiosk mode seemingly doesnt work under Wayland because it
cannot request to be active, so to work around this, using
Westons kiosk-shell and then disable_context_menu config, gives
sort of the same idea.